+++ src/components/common/StarToggleIcon.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { useColorMode, useDisclosure, useToken } from '@chakra-ui/react';
-import { FiStar } from 'react-icons/fi';
+import { FiMinusCircle, FiStar } from 'react-icons/fi';
 import useViewer from '@hooks/queries/useViewer';
 import useUpdateStarredProfiles from '@hooks/mutations/useUpdateStarredProfiles';
 import TooltipIconButton from '@common/inputs/TooltipIconButton';
@@ -14,6 +14,8 @@ interface Props {
 export default function StarToggleIcon({ id, ...props }: Props) {
 	const { loggedInId, starredProfiles } = useViewer();
 	const [isStarred, setIsStarred] = useState<boolean>(false);
+	const [hovered, setHovered] = useState<boolean>(false);
+
 	const {
 		updateStarredProfilesMutation,
 		results: { loading },
@@ -22,50 +24,60 @@ export default function StarToggleIcon({ id, ...props }: Props) {
 	const { colorMode } = useColorMode();
 	const [orange, lightGray, darkGray] = useToken('colors', ['orange.300', 'gray.300', 'gray.600']);
 
-	const { isOpen: isOpenConfirm, onOpen: onOpenConfirm, onClose: onCloseConfirm } = useDisclosure();
+	const {
+		isOpen: isOpenConfirmation,
+		onOpen: onOpenConfirmation,
+		onClose: onCloseConfirmation,
+	} = useDisclosure();
 
+	// Sets whether or not the profile is starred.
 	useEffect(() => {
 		if (starredProfiles) {
 			setIsStarred(starredProfiles.includes(id));
 		}
-	}, [starredProfiles, id]);
+
+		return () => setIsStarred(false);
+	}, [JSON.stringify(starredProfiles), id]);
 
 	const updateStarredProfilesHandler = () => {
 		if (!id) return;
 
-		// Optimistically update local state
-		setIsStarred(!isStarred);
-
 		// Fire the mutation
-		updateStarredProfilesMutation(id);
+		updateStarredProfilesMutation(id).then(() => {
+			onCloseConfirmation();
+		});
 	};
-
-	const iconLabel = isStarred ? 'Unstar this profile' : 'Star this profile';
 
 	return (
 		<>
 			<TooltipIconButton
 				icon={
-					<FiStar
-						color={isStarred ? orange : ''}
-						fill={isStarred ? orange : 'transparent'}
-						stroke={colorMode === 'dark' ? lightGray : darkGray}
-						size={20}
-					/>
+					hovered && isStarred ? (
+						<FiMinusCircle fill={orange} />
+					) : (
+						<FiStar
+							color={isStarred ? orange : ''}
+							fill={isStarred ? orange : 'transparent'}
+							stroke={colorMode === 'dark' ? lightGray : darkGray}
+							size={20}
+						/>
+					)
 				}
 				cursor='pointer'
 				borderRadius='full'
-				label={iconLabel}
-				onClick={isStarred ? onOpenConfirm : updateStarredProfilesHandler}
+				label={isStarred ? 'Unstar this profile' : 'Star this profile'}
+				onClick={isStarred ? onOpenConfirmation : updateStarredProfilesHandler}
 				mx={2}
 				isLoading={loading}
 				isDisabled={!loggedInId}
+				onMouseEnter={() => setHovered(true)}
+				onMouseLeave={() => setHovered(false)}
 				{...props}
 			/>
 
 			<ConfirmActionDialog
-				isOpen={isOpenConfirm}
-				onClose={onCloseConfirm}
+				isOpen={isOpenConfirmation}
+				onClose={onCloseConfirmation}
 				confirmAction={updateStarredProfilesHandler}
 				headerText='Unstar this profile?'
 				buttonsText={{ confirm: 'Yes', cancel: 'No' }}